refactor(api): migrate chat.postMessage endpoint to new API.v1 format
