+++ .changeset/migrate-chat-postMessage-to-AJV.md
@@ -1,0 +1,6 @@
+---
+'@rocket.chat/rest-typings': minor
+'@rocket.chat/meteor': minor
+---
+
+Migrate chat.postMessage endpoint to make it openAPI and AJV compatible
