+++ packages/rest-typings/src/v1/chat.ts
@@ -739,132 +739,6 @@ const ChatGetDeletedMessagesSchema = {
 
 export const isChatGetDeletedMessagesProps = ajv.compile<ChatGetDeletedMessages>(ChatGetDeletedMessagesSchema);
 
-type ChatPostMessage =
-	| {
-			roomId: string | string[];
-			text?: string;
-			alias?: string;
-			emoji?: string;
-			avatar?: string;
-			attachments?: MessageAttachment[];
-			customFields?: IMessage['customFields'];
-	  }
-	| {
-			channel: string | string[];
-			text?: string;
-			alias?: string;
-			emoji?: string;
-			avatar?: string;
-			attachments?: MessageAttachment[];
-			customFields?: IMessage['customFields'];
-	  };
-
-const ChatPostMessageSchema = {
-	oneOf: [
-		{
-			type: 'object',
-			properties: {
-				roomId: {
-					oneOf: [
-						{ type: 'string' },
-						{
-							type: 'array',
-							items: {
-								type: 'string',
-							},
-						},
-					],
-				},
-				text: {
-					type: 'string',
-					nullable: true,
-				},
-				alias: {
-					type: 'string',
-					nullable: true,
-				},
-				emoji: {
-					type: 'string',
-					nullable: true,
-				},
-				avatar: {
-					type: 'string',
-					nullable: true,
-				},
-				attachments: {
-					type: 'array',
-					items: {
-						type: 'object',
-					},
-					nullable: true,
-				},
-				tmid: {
-					type: 'string',
-				},
-				customFields: {
-					type: 'object',
-					nullable: true,
-				},
-				parseUrls: {
-					type: 'boolean',
-				},
-			},
-			required: ['roomId'],
-			additionalProperties: false,
-		},
-		{
-			type: 'object',
-			properties: {
-				channel: {
-					oneOf: [
-						{ type: 'string' },
-						{
-							type: 'array',
-							items: {
-								type: 'string',
-							},
-						},
-					],
-				},
-				text: {
-					type: 'string',
-					nullable: true,
-				},
-				alias: {
-					type: 'string',
-					nullable: true,
-				},
-				emoji: {
-					type: 'string',
-					nullable: true,
-				},
-				avatar: {
-					type: 'string',
-					nullable: true,
-				},
-				attachments: {
-					type: 'array',
-					items: {
-						type: 'object',
-					},
-					nullable: true,
-				},
-				customFields: {
-					type: 'object',
-					nullable: true,
-				},
-				parseUrls: {
-					type: 'boolean',
-				},
-			},
-			required: ['channel'],
-			additionalProperties: false,
-		},
-	],
-};
-
-export const isChatPostMessageProps = ajv.compile<ChatPostMessage>(ChatPostMessageSchema);
-
 type ChatGetURLPreview = {
 	roomId: IRoom['_id'];
 	url: string;
@@ -980,13 +854,6 @@ export type ChatEndpoints = {
 					previous: string | null;
 				};
 			};
-		};
-	};
-	'/v1/chat.postMessage': {
-		POST: (params: ChatPostMessage) => {
-			ts: number;
-			channel: IRoom;
-			message: IMessage;
 		};
 	};
 	'/v1/chat.syncThreadMessages': {
